pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.12"
dependencies = [
    "python-dotenv>=1.2.2",
    "vision-agents[gemini,getstream,redis]>=0.5.4",
]
